pyproject.toml
[project]
name = "djangorest-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Acejkee",email = "belousow.ilja@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (==4.2)",
    "djangorestframework (==3.14.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-autoslug (==1.9.9)",
    "pytils (==0.4.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "python-dotenv (==1.0.1)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-celery-email (>=3.0.0,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
setuptools = "^75.8.0"

